refactor: move transbank api request class to utils
